tests!(scipt/run_before_test, conftest): will use pytest_configure
